Broaden install targets to include VS15 Preview
While currently still in a fairly early preview stage, preview 4 to be exact, VS15 is already quite polished and offers some extra niceties for TypeScript projects including opening a folder as a solution. This change allows the extension to be discovered and installed into any version of VS15. Thank you.